refactor: move provider configs to central place
