+++ scripts/generateWidgetUrl.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 import yargs from 'yargs'
-import { ProviderIds } from '../src/types'
 import { CryptoType, FiatType } from '@fiatconnect/fiatconnect-types'
+import { ProviderIds } from '../src/providerConfig'
 
 function loadConfig() {
   return yargs
