+++ src/components/SignInScreen.tsx
@@ -11,8 +11,8 @@ import { useAccount } from 'wagmi'
 import styled from 'styled-components'
 import { getLinkedAccount, getKycStatus } from '../FiatConnectClient'
 import { useFiatConnectConfig } from '../hooks'
-import { providerIdToProviderName } from '../constants'
 import { QueryParams } from '../schema'
+import { providerIdToProviderName } from '../providerConfig'
 
 interface Props {
   onError: (title: string, message: string) => void
